refactor(runner,cli): relocate the relaxation helpers beside the validator (D6)
Pure move, no behavior change. relaxDefaultedRequired and localRefTarget
re-implemented the runtime's default-satisfaction rule inside
packages/cli/lib/callable.ts, and C5 (closed-world verb inputs enforced
at dispatch) needs the identical relaxation server-side — two copies of
"what does a default satisfy" is how the CLI and the runtime drift
apart. Both helpers now live verbatim beside validateSchemaValue in
packages/runner/src/cfc/schema-sanitization.ts, exported from the
defining module and imported directly by the CLI through a new
./cfc/schema-sanitization package export — no re-export shim, nothing
added to cfc/mod.ts.

The relaxation unit tests move with the code
(packages/runner/test/cfc-defaulted-required-relaxation.test.ts),
pinned against the same relax-then-validate composition the gate
applies; the CLI keeps the tests that exercise the gate itself
(refusal at both entry points, id not spent) plus one
defaulted-required pin proving the gate applies the relaxation.

The helper's doc comment now names the boundary — what relaxation does
NOT check: only `required` lists are rewritten, so additionalProperties,
patternProperties, minProperties and the other constraints, const/enum,
not and if/then/else, and oneOf exclusivity are all judged against the
original schema text. Each is a potential refused-but-valid call when a
generated schema leans on a defaulted property through them.
