+++ packages/runner/deno.jsonc
@@ -35,6 +35,7 @@
     "./toolshed-http-auth": "./src/toolshed-http-auth.ts",
     "./cfc": "./src/cfc/mod.ts",
     "./cfc/label-view-core": "./src/cfc/label-view-core.ts",
-    "./cfc/migration-reason": "./src/cfc/migration-reason.ts"
+    "./cfc/migration-reason": "./src/cfc/migration-reason.ts",
+    "./cfc/schema-sanitization": "./src/cfc/schema-sanitization.ts"
   }
 }
