+++ packages/runner/src/cfc/schema-sanitization.ts
@@ -1261,6 +1261,153 @@ const unmarkSchemaValueActive = (
   );
 };
 
+const isSchemaObject = (
+  schema: JSONSchema | undefined,
+): schema is Record<string, unknown> =>
+  typeof schema === "object" && schema !== null && !Array.isArray(schema);
+
+/**
+ * Follow local `#/$defs/...` / `#/definitions/...` references to the schema
+ * they name, so a `default` behind one is still seen. Chains are followed to
+ * their end; anything unresolvable (a remote ref, a missing entry, a cycle)
+ * yields the last schema reached rather than failing.
+ */
+export function localRefTarget(
+  schema: JSONSchema,
+  root: JSONSchema,
+): JSONSchema {
+  let current = schema;
+  const visited = new Set<object>();
+  while (isSchemaObject(current)) {
+    const ref = current.$ref;
+    if (typeof ref !== "string" || !isSchemaObject(root)) return current;
+    if (visited.has(current)) return current;
+    visited.add(current);
+    // `$defs` only, deliberately — NOT `definitions`. Hoisting emits `$defs`
+    // and `#/$defs/...` (schema-generator AGENTS.md: "anything that says
+    // `definitions` is out of date"), so a `definitions` ref is one the
+    // runtime cannot resolve either. Following it here would relax a required
+    // field on the strength of a default that never gets injected, letting an
+    // invalid payload through and spending its invocation id on a handling
+    // that receives no event — the exact failure the pre-dispatch gate exists
+    // to stop. Unresolvable refs keep the field required, so the call is
+    // refused before dispatch and the id survives.
+    const match = /^#\/\$defs\/(.+)$/.exec(ref);
+    if (!match) return current;
+    const pool = (root as Record<string, unknown>).$defs;
+    if (!isSchemaObject(pool as JSONSchema)) return current;
+    const target = (pool as Record<string, JSONSchema>)[match[1]];
+    if (target === undefined) return current;
+    current = target;
+  }
+  return current;
+}
+
+/**
+ * A copy of `schema` whose `required` lists omit properties that carry their
+ * own `default`.
+ *
+ * The runtime injects a property's default when the payload leaves it out (the
+ * schema read path in runner `schema.ts`), so such a property is satisfiable
+ * without the caller supplying it. Validating against the unrelaxed schema
+ * would reject payloads the verb would have accepted.
+ *
+ * This is honest only for a payload that is PRESENT (measured 2026-07-30,
+ * recorded on #5147): `SchemaObjectTraverser.traverseObjectWithSchema` (runner
+ * `traverse.ts`) fills each missing defaulted property of a present object
+ * before checking `required`, while a wholly absent event bypasses the object
+ * branch entirely — the handler sees `undefined` and no default is ever
+ * conjured. The CLI's absent-payload gate therefore normalizes an absent
+ * payload to `{}` before consulting this relaxation; absence is never excused
+ * by it.
+ *
+ * What this relaxation does NOT check — the boundary, named so nobody assumes
+ * it: only `required` lists are rewritten. Every other validation
+ * `validateSchemaValue` applies runs against the original schema text —
+ * `additionalProperties`, `patternProperties`, `minProperties` and the other
+ * object/array/string/number constraints, `const`/`enum`, `not` and
+ * `if`/`then`/`else`, and `oneOf` exclusivity. None of them is re-judged as
+ * if the runtime's defaults had already been filled in, so a schema that
+ * leans on a defaulted property through one of them (say `minProperties: 1`
+ * over a single all-defaulted property, or an `if` conditioned on it) can
+ * refuse a payload the runtime would have completed from defaults — a
+ * refused-but-valid call, the conservative side of this gate.
+ *
+ * `seen` both memoizes and breaks reference cycles: the relaxed copy is
+ * registered before its children are filled in, so a schema that reaches
+ * itself resolves to the copy already under construction.
+ */
+export function relaxDefaultedRequired(
+  schema: JSONSchema,
+  root: JSONSchema,
+  seen: Map<object, JSONSchema>,
+): JSONSchema {
+  if (!isSchemaObject(schema)) return schema;
+  const cached = seen.get(schema);
+  if (cached !== undefined) return cached;
+
+  const relaxed: Record<string, unknown> = { ...schema };
+  seen.set(schema, relaxed as JSONSchema);
+
+  const properties = schema.properties;
+  if (isSchemaObject(properties)) {
+    const defaulted = new Set<string>();
+    const next: Record<string, JSONSchema> = {};
+    for (
+      const [key, propSchema] of Object.entries(
+        properties as Record<string, JSONSchema>,
+      )
+    ) {
+      const target = localRefTarget(propSchema, root);
+      if (isSchemaObject(target) && target.default !== undefined) {
+        defaulted.add(key);
+      }
+      next[key] = relaxDefaultedRequired(propSchema, root, seen);
+    }
+    relaxed.properties = next;
+    if (Array.isArray(schema.required)) {
+      relaxed.required = (schema.required as string[]).filter(
+        (key) => !defaulted.has(key),
+      );
+    }
+  }
+
+  // `items` is a single schema here; the validator rejects the tuple form
+  // outright ("schema must be an object or boolean").
+  if (schema.items !== undefined) {
+    relaxed.items = relaxDefaultedRequired(
+      schema.items as JSONSchema,
+      root,
+      seen,
+    );
+  }
+
+  const fields = schema as Record<string, unknown>;
+  for (const combinator of ["anyOf", "oneOf", "allOf"]) {
+    const branches = fields[combinator];
+    if (Array.isArray(branches)) {
+      relaxed[combinator] = (branches as JSONSchema[]).map((entry) =>
+        relaxDefaultedRequired(entry, root, seen)
+      );
+    }
+  }
+
+  for (const pool of ["$defs", "definitions"]) {
+    const defs = fields[pool];
+    if (isSchemaObject(defs as JSONSchema)) {
+      const next: Record<string, JSONSchema> = {};
+      for (
+        const [key, entry] of Object.entries(defs as Record<string, JSONSchema>)
+      ) {
+        next[key] = relaxDefaultedRequired(entry, root, seen);
+      }
+      relaxed[pool] = next;
+    }
+  }
+
+  return relaxed as JSONSchema;
+}
+
 export const validateSchemaValue = (
   schema: JSONSchema,
   value: unknown,
